pyproject.toml
[tool.poetry]
name = "density-annotator"
version = "0.1.0"
description = "An annotation tool that creates a density map from an image."
authors = ["kenya-sk"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
hydra-core = "^1.1.1"
numpy = "^1.21.5"
opencv-python = "^4.5.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
